# scim-server/src/main/resources/gluu-scim-openapi.yaml
openapi: 3.0.1
info:
  title: SCIM API
  description: |
    Gluu SCIM 2.0 server API. Developers can think of SCIM as a REST API with endpoints exposing CRUD functionality
    (create, update, retrieve and delete) for identity management resources such as users, groups, and fido devices.
  license:
    name: Apache 2
    url: https://github.com/GluuFederation/scim/blob/master/LICENSE
  version: 4.3.0
servers:
- url: https://your.gluu.server.com/identity/restv1/scim/v2
...
paths:
  /Groups:
    get:
      tags:
      - group
      operationId: get-groups
      description: Query Group resources (see section 3.4.2 of RFC 7644)
      security:
        - scim_oauth: [https://gluu.org/scim/groups.read]
      parameters:
      - name: attributes
        in: query
        description: A comma-separated list of attribute names to return in the response
        schema:
          type: string
      - name: excludedAttributes
        in: query
        description: When specified, the response will contain a default set of attributes minus those listed here (as a comma-separated list)
        schema:
          type: string
      - name: filter
        in: query
        description: An expression specifying the search criteria. See section 3.4.2.2 of RFC 7644
        schema:
          type: string
        example: displayName co "Audit"
      - name: startIndex
        in: query
        description: The 1-based index of the first query result
        schema:
          type: integer
      - name: count
        in: query
        description: Specifies the desired maximum number of query results per page
        schema:
          type: integer
      - name: sortBy
        in: query
        description: The attribute whose value will be used to order the returned responses
        schema:
          type: string
      - name: sortOrder
        in: query
        description: Order in which the sortBy param is applied. Allowed values are "ascending" and "descending"
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        400:
          description: Parameter count exceeds the maximum allowed value or the filter
            supplied was unparsable
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: There was an unexpected failure executing the operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - group
      operationId: create-group
      description: Allows creating a Group resource via POST (see section 3.3 of RFC
        7644)
      security:
        - scim_oauth: [https://gluu.org/scim/groups.write]
      parameters:
      - name: attributes
        in: query
        description: A comma-separated list of attribute names to return in the response
        schema:
          type: string
      - name: excludedAttributes
        in: query
        description: When specified, the response will contain a default set of attributes minus those listed here (as a comma-separated list)
        schema:
          type: string
      requestBody:
        description: Payload that represents the Group to create
        content:
          application/scim+json:
            schema:
              $ref: '#/components/schemas/GroupResource'
          application/json:
            schema:
              $ref: '#/components/schemas/GroupResource'
        required: true
      responses:
        201:
          description: Successful operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/GroupResource'
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResource'
        400:
          description: An invalid value was passed in the payload
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: There is a conflict with an already existing group. Uniqueness
            is assumed over displayName
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: There was an unexpected failure executing the operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: group
...
components:
  schemas:
...
    ErrorResponse:
      required:
      - status
      type: object
      properties:
        schemas:
          type: array
          items:
            type: string
            example: urn:ietf:params:scim:api:messages:2.0:Error
        status:
          type: string
          description: HTTP status code as string
        scimType:
          type: string
          description: A detail error keyword. See table 9 of RFC 7644
        detail:
          type: string
          description: A detailed human-readable message of the error
      description: See section 3.12 of RFC 7644
...
    BasicListResponse:
      type: object
      properties:
        schemas:
          type: array
          items:
            type: string
            example: urn:ietf:params:scim:api:messages:2.0:ListResponse
        totalResults:
          type: integer
          description: Total number of results returned by the search. The value may be larger than the number of resources returned due to pagination
        startIndex:
          type: integer
          description: The 1-based index of the first result in the current set of search results
        itemsPerPage:
          type: integer
          description: The number of resources returned in a results page
    GroupListResponse:
      description: Results for groups search. See section 3.4.2.4 of RFC 7644
      allOf:
      - $ref: '#/components/schemas/BasicListResponse'
      - type: object
      - type: object
        properties:
          Resources:
            type: array
            items:
              $ref: '#/components/schemas/GroupResource'
...
    Meta:
      type: object
      properties:
        resourceType:
          type: string
        created:
          type: string
        lastModified:
          type: string
        location:
          type: string
      description: Descriptive information about a resource. See section 3.1 of RFC 7643
    Member:
      type: object
      description: Represents a member of a Group resource
      properties:
        $ref:
          type: string
          description: URI of the SCIM resource
        type:
          type: string
          description: The type of member. Only "User" is allowed
        display:
          type: string
          description: A human readable name, primarily used for display purposes
        value:
          type: string
          description: Identifier (ID) of the resource
...
    BaseResource:
      type: object
      properties:
        schemas:
          type: array
          description: URIs that are used to indicate the namespaces of the SCIM schemas that define the attributes present in the current structure
          items:
            type: string
        id:
          type: string
          description: A unique identifier for a SCIM resource. See section 3.1 of RFC 7643
        meta:
          $ref: '#/components/schemas/Meta'
    GroupResource:
      description: Represents a group resource. See section 4.2 of RFC 7643
      allOf:
      - $ref: '#/components/schemas/BaseResource'
      - type: object
      - type: object
        properties:
          displayName:
            type: string
            description: Group name suitable for display to end-users
          members:
            type: array
            items:
              $ref: '#/components/schemas/Member'
...
  securitySchemes:
    scim_oauth:
      type: oauth2
      description: |
        Endpoints protected by a bearer token passed in the Authorization header. 
        It is assumed gluu-scim is configured to operate in oauth protection mode. Other protection
        modes may be available, please check SCIM docs
      flows:
        clientCredentials:
          tokenUrl: 'https://your.gluu.server.com/oxauth/restv1/token'
          scopes:
            https://gluu.org/scim/users.read: Query user resources
            https://gluu.org/scim/users.write: Modify user resources
            https://gluu.org/scim/groups.read: Query group resources
            https://gluu.org/scim/groups.write: Modify group resources
            https://gluu.org/scim/fido.read: Query fido resources
            https://gluu.org/scim/fido.write: Modify fido resources
            https://gluu.org/scim/fido2.read: Query fido 2 resources
            https://gluu.org/scim/fido2.write: Modify fido 2 resources
            https://gluu.org/scim/all-resources.search: Access the root .search endpoint 
            https://gluu.org/scim/bulk: Send requests to the bulk endpoint